Fix handling of links that only consist of fragment
If a link only consists of a fragment (#something), we must not resolve
that relative to the "base" url, which is basically the "path" of the
current document. It should be resolved wrt the document URL (including
the file name), which we currently don't have available in the container
implementation. Just return the fragment unchanged and let the link
handler take care of that, like it was before
ae2afed543ff2499aab6004394d69684b0911d53.

Amends ae2afed543ff2499aab6004394d69684b0911d53

Task-number: QTCREATORBUG-25978
Change-Id: Ibca5f653d107c463a82f538c86aaf607821111bb
Reviewed-by: Robert Löhning <robert.loehning@qt.io>